requirements.txt
azure-identity>=1.15.0
azure-monitor-query>=1.4.0
databricks-sql-connector>=2.12.0
